Con esto así como está me funciona la app bien, con lo número decimales.
Inclui la info que faltaba de los Deportes

+++ ADOL.APP/ADOL.APP.DataBase/dbo/Tables/ApuestasDeUsuarios.sql
@@ -1,12 +1,14 @@
 ﻿CREATE TABLE [dbo].[ApuestasDeUsuarios] (
-    [ID]                 INT          IDENTITY (1, 1) NOT NULL,
-    [Token]              VARCHAR (36) NOT NULL,
-    [Amount]             FLOAT NOT NULL,
-    [ApuestaDeportivaID] INT          NOT NULL,
-    [Acierto]            BIT          NULL,
-    [BetType] VARCHAR(50) NOT NULL, 
-    [BetPrice] FLOAT NOT NULL, 
+    [ID]                 INT            IDENTITY (1, 1) NOT NULL,
+    [Token]              VARCHAR (36)   NOT NULL,
+    [Amount]             DECIMAL (8, 2) NOT NULL,
+    [ApuestaDeportivaID] INT            NOT NULL,
+    [Acierto]            BIT            NULL,
+    [BetType]            VARCHAR (50)   NOT NULL,
+    [BetPrice]           DECIMAL (8, 2) NOT NULL,
     CONSTRAINT [PK_ApuestasDeUsuarios] PRIMARY KEY CLUSTERED ([ID] ASC),
     CONSTRAINT [ApuestasDeUsuarios_FK1] FOREIGN KEY ([ApuestaDeportivaID]) REFERENCES [dbo].[ApuestasDeportivas] ([ID])
 );
+
+
